refactor(ui): extract complex checkout components
- Extracted Payment Methods section into CheckoutPaymentMethods.tsx
- Extracted Offers Bottom Sheet into CheckoutOffersSheet.tsx
- Extracted Confirmation Modal into CheckoutConfirmModal.tsx
- Simplified Checkout.tsx by using the newly extracted components and removing unused states, hooks, and imports.
